Create Core Repository - Database Module Structure
Add core database & repository architecture.

Stock entity structure created to reflect the endpoint in StockWebService.
A fake repository and the desired database structure are created around the endpoint data.

This structure enforces the separation of concerns and defines the basics needs so that we can keep our code robust and testable, especially when there are multiple developers.